Change 2i to use long values
The original implementation of 2i used int for _int indexes
which is caused an upper bound of 2^31. The client will
now use long values eliminating this restriction.

When possible backwards compatability was maintained, specifically
regarding the @RiakIndex notation and the FetchIndex class. Users of
the IRiakObject interface calling getIntIndex() or allIntIndexes()
will need to change the type being returned from Set<Integer> to Set<Long>

This addresses #112

+++ src/main/java/com/basho/riak/client/DefaultRiakObject.java
@@ -410,7 +410,7 @@ public class DefaultRiakObject implements IRiakObject {
      * 
      * @see com.basho.riak.client.IRiakObject#allIntIndexes()
      */
-    public Map<IntIndex, Set<Integer>> allIntIndexes() {
+    public Map<IntIndex, Set<Long>> allIntIndexes() {
         return indexes.getIntIndexes();
     }
 
@@ -419,7 +419,7 @@ public class DefaultRiakObject implements IRiakObject {
      * 
      * @see com.basho.riak.client.IRiakObject#getIntIndex(java.lang.String)
      */
-    public Set<Integer> getIntIndex(String name) {
+    public Set<Long> getIntIndex(String name) {
         return indexes.getIntIndex(name);
     }
 
@@ -432,9 +432,9 @@ public class DefaultRiakObject implements IRiakObject {
     }
 
     /* (non-Javadoc)
-     * @see com.basho.riak.client.IRiakObject#addIndex(java.lang.String, int)
+     * @see com.basho.riak.client.IRiakObject#addIndex(java.lang.String, long)
      */
-    public IRiakObject addIndex(String index, int value) {
+    public IRiakObject addIndex(String index, long value) {
         indexes.add(index, value);
         return this;
     }
